Add validation for SWIFT code and country ISO2 in request and response handling

+++ src/main/java/org/example/swiftapi/SwiftInfoController.java
@@ -24,6 +24,9 @@ public class SwiftInfoController {
     public ResponseEntity<SwiftInfoUniversalResponseDTO> getSwiftInfo(
             @PathVariable("swift-code") String swiftCode
     ) {
+        if (swiftCode.length() != 11) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
         Optional<SwiftInfoUniversalResponseDTO> responseDTO =
                 swiftInfoService.findSwiftInfo(swiftCode);
 
@@ -35,6 +38,9 @@ public class SwiftInfoController {
     public ResponseEntity<SwiftInfoCountryResponseDTO> getCountrySwiftInfo(
             @PathVariable("countryISO2code") String countryISO2code
     ) {
+        if (countryISO2code.length() != 2) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
         Optional<SwiftInfoCountryResponseDTO> responseDTO =
                 swiftInfoService.findSwiftInfosRelatedWithCountry(countryISO2code);
         return responseDTO.map(ResponseEntity::ok)
@@ -45,6 +51,9 @@ public class SwiftInfoController {
     public ResponseEntity<OnlyMessageResponseDTO> createSwiftInfo(
             @RequestBody AddSwiftInfoRequestDTO addSwiftInfoRequestDTO
     ) {
+        if (!addSwiftInfoRequestDTO.isCorrectSwiftInfo()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
         String response = swiftInfoService.createSwiftInfo(addSwiftInfoRequestDTO);
         OnlyMessageResponseDTO responseDTO = new OnlyMessageResponseDTO();
         responseDTO.setMessage(response);
@@ -53,6 +62,9 @@ public class SwiftInfoController {
 
     @DeleteMapping("/{swift-code}")
     public ResponseEntity<OnlyMessageResponseDTO> deleteSwiftInfo(@PathVariable("swift-code") String swiftCode) {
+        if (swiftCode.length() != 11) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
         String response = swiftInfoService.deleteSwiftInfo(swiftCode);
         OnlyMessageResponseDTO responseDTO = new OnlyMessageResponseDTO();
         responseDTO.setMessage(response);
